Writing Tests using unittest library
